+++ bakervpr/src/main/java/com/baker/sdk/vpr/bean/request/VprMatchRequest.kt
@@ -11,7 +11,7 @@ data class VprMatchRequest(
     override val access_token: String,
     override val audio: ByteArray,
     override val format: String,
-    override val scoreThreshold: Double,
+    override val scoreThreshold: Float,
     /**调用创建声纹库接口返回的 id*/
     val matchId: String
 ) : VprBaseRequest()
